Add files via upload
Summary
After the COVID-19 vaccine was developed at the latter end of 2020, there have been several questions on the minds of individuals on issues like the speed at which this vaccine was developed, the extent of research and knowledge of the potency of the vaccine, and even conspiracy theories on the agenda of the government.
In this report, I will be exploring the adverse effects that have been recorded so far from individuals who have been vaccinated within the United States, to answer questions like: What is the vaccination rate in the U.S by the state?  What is the probability of a vaccinated individual getting an adverse reaction from each of the vaccines? What kinds of adverse effects have vaccinated individuals experienced? What is the probability of adverse reaction from the vaccine when you have an underlying illness or allergies?

Today, the highly transmissible Delta variant of the COVID-19 virus is the dominant strain in the United States, representing 51% of new cases. Studies show that the three major vaccines are effective at preventing severe illness from COVID-19, Pfizer – 95%, Moderna – 94.1%, and Johnson & Johnson – 72% efficacy. These values are still being studied for the Delta variant but each vaccine has proven to prevent severe illness from the virus regardless of the variant.

Problem Statement
According to a Gallup survey conducted in March and April, 23% of respondents who were hesitant to receive the COVID-19 vaccine were wanting to confirm that the vaccine was safe, 16% expressed concern about the timeline for the development of the vaccine, another 16% have a mistrust for vaccines in general, 10% said they already have immunity from contracting the virus and 10% cited existing allergies or concerns about allergies as the reason they do not plan to get vaccinated. 

Problems that Occurred
1.	Dataset Issues: issues with obtaining a comprehensive dataset with the required information, datasets were fragmented based on different factors relating to the topic and required extensive data modeling.
2.	Load Editor Error: this error caused the project to shut off and remain irretrievable, the app required troubleshooting to restore. A backup creation plan was developed as a lesson learned to mitigate future incidents.
3.	Data Issues: data was raw/dirty and required cleaning and expressions to eliminate unwanted values.

Findings and Suggestions
1.	The rewards of getting vaccinated outweighs the risk. There is 0.1% of an adverse reaction which is low in comparison to the probability of death from COVID-19 1.8%.  
2.	People who have pre-existing conditions such as asthma, hypertension, diabetes, and hypothyroidism should discuss with their health care providers and make plans for check-in post-vaccination.
3.	A person with asthma who is worried about adverse reactions should consider getting the Johnson and Johnson vaccine over Pfizer and Moderna.
4.	Regardless of vaccine type, hypertension as a pre-existing condition is a major factor in getting serious adverse reactions from the COVID-19 vaccine.
5.	Dizziness and anxiety are the two most common first symptoms after vaccination.
6.	Individuals in the Northeast are more likely to get vaccinated or have easier access to vaccines than people in Idaho and the Southeast.